整理代码

+++ src/main/java/com/pixivdownloader/core/entity/novel/NovelRanking.java
@@ -10,6 +10,6 @@ import lombok.Data;
 public class NovelRanking {
     private Integer novelId;
     private Integer id;
-    private Integer bookmarkId;
+    private Long bookmarkId;
     private Integer rank;
 }
